Added ``apt_preferences__preset_list`` for advanced users.
The presets are intended for advanced users who want to install software
from different releases but want to keep there main system on the stable
release.

When enabling these presets, you will need to set explicit preferences for
packages you want to install from different releases or from backports.

Configuring this in APT is needed because releases (stable, testing,
unstable) have by default the same priority 500 so when including different
releases, APT would just select the newest version (usually unstable). To
avoid this, APT preferences are used to change the preferences to prefer
stable over testing over unstable.

@drybjed and me discoursed a bit how to add such examples to the role.
The previous consent was to add it to the documentation but I think
adding it to the defaults has some benefits in this case which are that
users can directly use the presets and benefit from updates directly.
Also, this avoids creating redundancy.

The preferences are tested by me for some time now.